fixup! Fixes prepare.sh
